requirements.txt

tensorflow-gpu 
numpy 
opencv-python
matplotlib
pandas


